Refactoring: CSS > SASS #2
