+++ src/styles/components/ChallengeBox.module.scss
@@ -1,0 +1,110 @@
+.challengeBoxContainer {
+    height: 100%;
+
+    background: var(--white);
+    box-shadow: 0 0 60px rgba(0, 0, 0, 0.05);
+    border-radius: 5px;
+
+    padding: 1.5rem 2.5rem;
+
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+}
+
+.challengeNotActive {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    strong {
+        font-size: 1.5rem;
+        font-weight: 500;
+        line-height: 1.4;
+    }
+
+    p {
+        max-width: 70%;
+        margin-top: 3rem;
+
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        line-height: 1.4;
+
+        img {
+            margin-bottom: 1.5rem;
+        }
+    }
+}
+
+.challengeActive {
+    height: 100%;
+
+    display: flex;
+    flex-direction: column;
+
+    header {
+        color: var(--blue);
+        font-weight: 600;
+        font-size: 1.25rem;
+        padding: 0 2rem 1.5rem;
+        border-bottom: 1px solid var(--gray-line);
+    }
+
+    main {
+        flex: 1;
+
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+
+        strong {
+            font-weight: 600;
+            font-size: 2rem;
+            color: var(--title);
+            margin: 1.5rem 0 1rem;
+        }
+
+        p {
+            line-height: 1.5;
+        }
+    }
+
+    footer {
+        display: grid;
+        grid-template-columns: 1fr 1fr;
+        gap: 1rem;
+
+        button {
+            height: 3rem;
+
+            display: flex;
+            align-items: center;
+            justify-content: center;
+            border: 0;
+            outline: none;
+            border-radius: 5px;
+
+            color: var(--white);
+            font-size: 1rem;
+            font-weight: 600;
+            transition: filter 0.2s;
+
+            &.challengeFailedButton {
+                background: var(--red);
+            }
+
+            &.challengeSucceceedButton {
+                background: var(--green);
+            }
+
+            &:hover {
+                filter: brightness(0.9);
+            }
+        }
+    }
+}
