+++ src/styles/components/CompletedChallenges.module.scss
@@ -6,14 +6,13 @@
     margin: 3.5rem 0;
     padding-bottom: 1rem;
     border-bottom: 1px solid #d7d8da;
-
     font-weight: 500;
-}
 
-.completedChallengesContainer span:first-child{
-    font-size: 1.25rem;
-}
+    span:first-child{
+        font-size: 1.25rem;
+    }
 
-.completedChallengesContainer span:last-child{
-    font-size: 1.5rem;
+    span:last-child{
+        font-size: 1.5rem;
+    }
 }
